tests/wpt/tests/workers/interfaces/WorkerUtils/WindowTimers/005.html

<!doctype html>
<title>setInterval when closing</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('005.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data, 1);
    this.done();
  });
});
</script>
